Add --update option to abs2osx.
As discussed on the listserv, change the --checkout method to actually be an export and have abs2osx handle the --update itself.

To handle abs2osx related conflicts, the files are converted prior to diff3 being run on them. This leaves us open to merge conflicts if abs2osx is updated, however it means most of the time resolving them is easier as the paths are consistently in /opt/arch, rather than a hybrid like if we reversed the abs2osx before merging.

Conflicts are handled the same way they are in git/svn/etc, that the merged file contains conflict markers and the user must resolve these manually.